requirements.txt
tqdm~=4.56.0
torch~=1.7.1
numpy~=1.19.5
scikit-image~=0.18.1
matplotlib~=3.3.3
Pillow~=8.1.0
opencv-python~=4.5.1.48
torchvision~=0.8.2
scipy~=1.6.0